+++ packages/react-craft-ai-decision-tree/src/components/nodeInformation/nodeInformation.jsx
@@ -41,6 +41,7 @@ const CloseButtonDiv = styled('div')`
 
 const NodeInformation = ({
   closeNodeInformation,
+  minifyTree,
   selectedNodePath,
   style,
   tree,
@@ -56,6 +57,8 @@ const NodeInformation = ({
     const { distribution } = interpreter.getPrediction(interpreter.dt);
 
     const { min, max } = distribution || { min: undefined, max: undefined };
+
+    const minifyLeaf = minifyTree ? () => minifyTree(selectedNodePath) : null;
 
     return (
       <NodeInformationContainer className='node-informations' style={ style }>
@@ -80,6 +83,15 @@ const NodeInformation = ({
             interpreter={ interpreter }
             outputValues={ interpreter.dt.output_values }
           />
+          <span>
+            {
+              !dtNode.children && minifyTree ? (
+                <button onClick={ minifyLeaf }>
+                  Minify Tree
+                </button>
+              ) : null
+            }
+          </span>
         </InformationContainer>
       </NodeInformationContainer>
     );
@@ -95,6 +107,7 @@ NodeInformation.defaultProps = {
 
 NodeInformation.propTypes = {
   closeNodeInformation: PropTypes.func,
+  minifyTree: PropTypes.func,
   tree: PropTypes.object.isRequired,
   selectedNodePath: PropTypes.string.isRequired,
   width: PropTypes.number,